requirements.txt
numpy==1.26.4
opencv-python-headless==4.9.0.80
pillow==10.0.0
segment-anything==1.0
torch==2.1.0
torchvision==0.16.0
runpod==1.8.1
